fix(rust): erros de compilação do cargo check (rodada 2, via anotações do CI)
- anigo-vrm: dependência thiserror faltando no Cargo.toml
- exporter.rs: duplo &mut bin em push_view(..., bin.write_X(..)) — extrair o
  offset para local (13 locais); Value::from(track.path) → .clone()
- lib.rs: resolver de buffer devolve o buffer do índice (BadBufferIndex);
  From<Vrm1Error> for GltfError; import Value movido pro módulo de teste
- vrm1.rs: temporário &Vrm1MToon::default() com lifetime curto — nomear
- headless.rs: cache: None no RenderPipelineDescriptor (wgpu 24);
  depth_resolve_attachment → depth_resolve_target (nome do wgpu; o browser
  usa depthResolveAttachment); bind group DoF no padrão BindingResource
- uniforms.rs: face_sdf_enabled/eye_enabled bool via u32

+++ crates/anigo-renderer/src/uniforms.rs
@@ -137,14 +137,14 @@ impl From<&anigo_core::StylizedMaterial> for MaterialUniform {
             params7: [
                 m.face_shadow_offset,
                 m.face_shadow_smoothness,
-                m.face_sdf_enabled as f32,
+                (m.face_sdf_enabled as u32) as f32,
                 0.0,
             ],
             // Fase 2 (#43): olho anime (parallax + highlights)
             params8: [
                 m.eye_depth_scale,
                 m.eye_highlight_intensity,
-                m.eye_enabled as f32,
+                (m.eye_enabled as u32) as f32,
                 0.0,
             ],
         }
